Add GLFramebuffer, refactor FrameBuffer and FrameBufferCubempa to extend it
